share.riseup.net [unfinished]
Everything works except that the encryption is wrong:
Up1 wants to use aes-256-ccm, not aes-256-cbc, presumably
because authenticated encryption is stronger: if someone
could interject malicious encrypted text maybe they could do
..something to you? Anyway, it's a good baseline, probably,
and openssl *has* this algorithm buried inside it somewhere,
but it doesn't expose it via its cli.

http://www.fredriks.se/?p=23 has some C
code you can compile against OpenSSL that might work.

Grrrrrr.

At this point the best way to support riseup
might be to insist on using their (maintained) client:
https://0xacab.org/riseup/share/up1-cli-client-nodejs
but that's in node, and would therefore pull in megabytes
of cruft just to be able to run this.